requirements.txt
certifi==2016.9.26
future==0.16.0
lxml==3.7.2
python-telegram-bot==12.0.0b1
requests==2.12.4
urllib3==1.19.1
beautifulsoup4==4.8.1
pycryptodome==3.9.4